server/requirements.txt
Flask_RESTful==0.3.8
Flask==1.1.2
Flask_SQLAlchemy==2.4.4
python_frontmatter==0.5.0
